refactor: modularize spotlight
