+++ src/spotlight/mod.rs
@@ -1,52 +1,13 @@
-use crate::node::NodeID;
-
-pub mod parser;
 pub mod commands;
-pub mod render;
+pub mod input;
+pub mod parser;
+pub mod result;
+pub mod search;
 
-pub use commands::{COMMANDS, command_preview};
-
-pub enum SpotlightMode {
-    Command,
-    Node,
-    Tag,
-    Jump,
-}
-
-pub enum SpotlightResult {
-    Command(String),
-    Node(NodeID),
-    Tag(String),
-    Jump(NodeID),
-}
-
-pub struct Spotlight {
-    pub input_buffer: String,
-    pub result_list: Vec<SpotlightResult>,
-    pub focus_index: usize,
-    pub mode: SpotlightMode,
-}
-
-/// Compute a fuzzy match score. Delegates to [`parser::fuzzy_score`].
-fn fuzzy_score(candidate: &str, query: &str) -> Option<usize> {
-    parser::fuzzy_score(candidate, query)
-}
-
-/// Return command suggestions using the predictive ranking parser.
-pub fn command_suggestions(input: &str) -> Vec<&'static str> {
-    parser::rank(input.trim_start_matches('/'), &COMMANDS)
-        .into_iter()
-        .take(5)
-        .collect()
-}
-
-/// Return command suggestions along with fuzzy match scores.
-pub fn command_suggestions_scored(input: &str) -> Vec<(&'static str, usize)> {
-    parser::rank_with_scores(input.trim_start_matches('/'), &COMMANDS)
-        .into_iter()
-        .take(5)
-        .collect()
-}
+pub use commands::{command_preview, COMMANDS};
+pub use input::{Spotlight, SpotlightMode, SpotlightResult};
+pub use result::command_icon;
+pub use search::{command_suggestions, command_suggestions_scored, fuzzy_score};
 
 #[cfg(test)]
 mod tests {
