--- src/spotlight/render.rs
@@ -1,11 +1,0 @@
-pub fn command_icon(cmd: &str) -> &'static str {
-    match cmd {
-        "zen" => "📄",
-        "settings" => "⚙️",
-        "triage" => "🧭",
-        "gemx" => "💭",
-        "plugin" => "🔌",
-        "search" => "🔍",
-        _ => "🔍",
-    }
-}
